requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.3.0
dgl>=1.0.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.2.0
networkx>=2.8.0
matplotlib>=3.6.0
seaborn>=0.12.0
tqdm>=4.64.0
pyyaml>=6.0
loguru>=0.6.0

# Geospatial and data processing
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.4.0
osmnx>=1.5.0
geopy>=2.3.0
h3>=4.0.0

# Optimization and scientific computing
scipy>=1.9.0
cvxpy>=1.3.0
numba>=0.56.0

# Configuration and utilities
omegaconf>=2.2.0
hydra-core>=1.2.0
rich>=12.6.0
click>=8.1.0

# Optional: For visualization
plotly>=5.12.0
folium>=0.14.0
dash>=2.14.0
pyzmq>=25.0.0

# Development and testing
pytest>=7.2.0
black>=22.10.0
flake8>=5.0.0
mypy>=0.991

# Jupyter notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.20.0
